Generate a request for the APIs.guru API

Prepare a request for the APIs.guru API, using the opinionated framework defined in nectar::req_init(), nectar::req_modify(), nectar::req_tidy_policy(), and nectar::req_pagination_policy().

Usage

guru_req_prepare(
  path,
  query = list(),
  body = NULL,
  header = list(),
  cookie = list(),
  method = NULL,
  tidy_policy = nectar::tidy_policy_unknown(),
  call = rlang::caller_env()
)

Arguments

path

(character or list) The route to an API endpoint. Optionally, a list or character vector with the path as one or more unnamed arguments (which will be concatenated with "/") plus named arguments to glue::glue() into the path.

query

(character or list) An optional list or character vector of parameters to pass in the query portion of the request. Can also include a .multi argument to pass to httr2::req_url_query() to control how elements containing multiple values are handled.

body

(multiple types) An object to use as the body of the request. If any component of the body is a path, pass it through fs::path() or otherwise give it the class "fs_path" to indicate that it is a path.

header

(list or NULL) An optional list of headers to add to the request using httr2::req_headers(). A NULL value for an individual header will explicitly remove that header if it was previously set.

cookie

(list or NULL) An optional list of cookies to set on the request using httr2::req_cookies_set(). NULL elements are removed.

method

(length-1 character, optional) If the method is something other than GET or POST, supply it. Case is ignored.

tidy_policy

(nectar_tidy_policy or NULL) A tidying policy prepared with tidy_policy_prepare(). By default, tidy_policy_body_auto() is used to automatically apply resp_body_auto() to responses.

call

(environment) The environment from which a function was called, e.g. rlang::caller_env() (the default). The environment will be mentioned in error messages as the source of the error. This argument is particularly useful for functions that are intended to be called as utilities inside other functions.

Value

A httr2::request() object with additional class nectar_request.

Details

You may wish to export this function (if the API changes often or you do not fully implement the API, for example).